Implemented map view and progress indicator on loading data
[ADD] MapView to detail controller including custom map pin and
center/zoom of map to point
[ADD] Progress indicator while loading data via json request

+++ OpenVIE/DataEntryDetail.h
@@ -19,15 +19,15 @@
 - (id)initWithProperties:(NSString *)title
                    apiId:(NSString *)apiId 
              district:(int)district
-             coordX:(float)coordX
-             coordY:(float)coordY
+             coordX:(double)coordX
+             coordY:(double)coordY
          thumbImage:(UIImage *)thumbImage
           fullImage:(UIImage *)fullImage;
 
 - (id)initWithProperties:(NSString *)title
                    apiId:(NSString *)apiId
                 district:(int)district
-                  coordX:(float)coordX
-                  coordY:(float)coordY;
+                  coordX:(double)coordX
+                  coordY:(double)coordY;
 
 @end
